requirements.txt
argparse
datetime
icecream
moviepy
os
pickle
platform
pydub
request
shutil
srt
tkinter
whisper-timestamped